+++ lib/sequent/migrations/migrations.rb
@@ -7,3 +7,4 @@ end
 require_relative 'migrate_events'
 require_relative 'projectors'
 require_relative 'view_schema'
+require_relative 'functions'