+++ lib/sequent/sequent.rb
@@ -7,6 +7,7 @@ require_relative 'core/projector'
 require_relative 'core/workflow'
 require_relative 'core/value_object'
 require_relative 'generator'
+require_relative 'migrations/migrations'
 
 module Sequent
   def self.new_uuid