+++ spec/fixtures/db/1/item_records.sql
@@ -1,0 +1,7 @@
+CREATE TABLE item_records%SUFFIX% (
+    id serial NOT NULL,
+    aggregate_id uuid NOT NULL,
+    CONSTRAINT item_records_pkey%SUFFIX% PRIMARY KEY (id)
+);
+
+CREATE INDEX unique_aggregate_id%SUFFIX% ON item_records%SUFFIX% USING btree (aggregate_id);