+++ spec/spec_helper.rb
@@ -58,3 +58,14 @@ RSpec::Matchers.define :have_view_schema_table do |expected|
     %Q{expected view schema tables:\n  #{tables.join("\n  ")}\nto contain:\n  #{expected}}
   end
 end
+
+RSpec::Matchers.define :have_column do |expected|
+  match do |actual|
+    actual.reset_column_information
+    expect(actual.column_names).to include(expected)
+  end
+
+  failure_message do |actual|
+    %Q{expected table #{actual.table_name} to have column '#{expected}'}
+  end
+end