+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+# Changes since 3.3
+
+- Added [alter table](https://www.sequent.io/docs/concepts/migrations.html#2-altertable) capabilities to migrations. Useful for larger projections.
+- Added `Sequent::Projector.manages_no_tables`.
+
 # Changes since 3.2
 
 - Introduced `strict_check_attributes_on_apply_events`. Sequent will fail when calling `apply` with unknown attributes.
-
 
 # Changelog 3.2